Extract networking and strip out EndpointLegacy
