+++ trim/src/main/java/es/sandwatch/trim/Trim.java
@@ -1,16 +1,8 @@
 package es.sandwatch.trim;
 
-import es.sandwatch.trim.annotation.Endpoint;
-import org.apache.http.HttpResponse;
-import org.apache.http.client.HttpClient;
-import org.apache.http.client.methods.HttpGet;
-import org.apache.http.impl.client.HttpClientBuilder;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
 import java.lang.reflect.Field;
 import java.util.*;
 
@@ -48,7 +40,6 @@ public class Trim{
 
     private ApiSpecification specification;
     private ProgressListener listener;
-    private HttpClient client;
 
 
     /**
@@ -68,30 +59,17 @@ public class Trim{
      * @return the report object.
      */
     private @NotNull Report run(){
-        List<EndpointLegacy> endpoints = new ArrayList<>();
-        //Add all generic headers to all endpoints
-        for (Class<?> model:specification.getModels()){
-            Endpoint endpointAnnotation = model.getDeclaredAnnotation(Endpoint.class);
-            if (endpointAnnotation != null){
-                EndpointLegacy endpoint = new EndpointLegacy(endpointAnnotation.value(), model);
-                for (String header : specification.getHeaders().keySet()){
-                    endpoint.addHeader(header, specification.getHeaders().get(header));
-                }
-                endpoints.add(endpoint);
-            }
-        }
-
-        //Create the http client and the report objects
-        client = HttpClientBuilder.create().build();
+        //Create the fetcher and the report objects
+        Fetcher fetcher = new Fetcher(specification.getHeaders());
         Report report = new Report();
 
         //Initialize a counter
         int completed = 0;
 
         //Execute the requests to endpoints
-        for (EndpointLegacy endpoint:endpoints){
-            RequestResult result = getEndpointData(endpoint);
-            Report.EndpointReport endpointReport = report.addEndpointReport(endpoint, result);
+        for (Class<?> model:specification.getModels()){
+            Fetcher.RequestResult result = fetcher.getEndpointData(model);
+            Report.EndpointReport endpointReport = report.addEndpointReport(model, result);
 
             //If successful
             if (result.is2xx()){
@@ -102,7 +80,7 @@ public class Trim{
                 }
                 else {
                     //Create and populate the field list
-                    List<Parser.FieldNode<Field>> fields = Parser.parseClass(endpoint.getModel());
+                    List<Parser.FieldNode<Field>> fields = Parser.parseClass(model);
                     for (Parser.FieldNode field:fields){
                         //Determine if it exists in the API response
                         if (keys.contains(field.getName())){
@@ -119,177 +97,15 @@ public class Trim{
             }
 
             if (listener != null){
-                listener.onEndpointReportComplete(endpoint, ++completed);
+                listener.onEndpointReportComplete(model, ++completed);
             }
         }
 
         return report;
     }
 
-    /**
-     * Hits an endpoint and returns the result.
-     *
-     * @param endpoint the endpoint to hit.
-     * @return a bundle containing request code and result
-     */
-    private @NotNull RequestResult getEndpointData(EndpointLegacy endpoint){
-        //Create the request and add all the headers
-        HttpGet request = new HttpGet(endpoint.getUrl());
-        for (String header:endpoint.getHeaders().keySet()){
-            request.addHeader(header, endpoint.getHeaders().get(header));
-        }
-
-        RequestResult result = null;
-        BufferedReader reader = null;
-        try{
-            long startTime = System.currentTimeMillis();
-            //Execute the request and create the reader
-            HttpResponse response = client.execute(request);
-            reader = new BufferedReader(new InputStreamReader(response.getEntity().getContent()));
-
-            //Fetch the result
-            StringBuilder stringBuilder = new StringBuilder();
-            String line;
-            while ((line = reader.readLine()) != null){
-                stringBuilder.append(line);
-            }
-
-            float timeSecs = (System.currentTimeMillis() - startTime)/1000f;
-
-            //Create the result bundle
-            result = new RequestResult(timeSecs, response.getStatusLine().getStatusCode(), stringBuilder.toString());
-        }
-        catch (IOException iox){
-            iox.printStackTrace();
-        }
-        finally{
-            if (reader != null){
-                try{
-                    reader.close();
-                }
-                catch (IOException iox){
-                    iox.printStackTrace();
-                }
-            }
-        }
-
-        //If there is no result, something went south
-        if (result == null){
-            result = new RequestResult();
-        }
-        return result;
-    }
 
 
-    /**
-     * Class containing the relevant information about the result of an HTTP request.
-     *
-     * @author Ismael Alonso
-     * @version 1.0.0
-     */
-    class RequestResult{
-        private final float requestTime;
-        private final int statusCode;
-        private final String response;
-
-
-        /**
-         * Constructor. Call if the request failed.
-         */
-        private RequestResult(){
-            this(0F, -1, "Request failed");
-        }
-
-        /**
-         * Constructor. Call if the request got through to the server.
-         *
-         * @param requestTime the time that took to complete the request.
-         * @param statusCode the status code of the request.
-         * @param response the response to the request.
-         */
-        private RequestResult(float requestTime, int statusCode, @NotNull String response){
-            this.requestTime = requestTime;
-            this.statusCode = statusCode;
-            this.response = response;
-        }
-
-        /**
-         * Request time getter.
-         *
-         * @return the time that took to complete the request.
-         */
-        float getRequestTime(){
-            return requestTime;
-        }
-
-        /**
-         * Tells whether the request failed before it was sent.
-         *
-         * @return true if the request failed, false otherwise.
-         */
-        boolean requestFailed(){
-            return statusCode == -1;
-        }
-
-        /**
-         * Tells whether the request yielded a 2xx status code.
-         *
-         * @return true if the request yielded a 2xx status code, false otherwise.
-         */
-        boolean is2xx(){
-            return statusCode >= 200 && statusCode < 300;
-        }
-
-        /**
-         * Tells whether the request yielded a 3xx status code.
-         *
-         * @return true if the request yielded a 3xx status code, false otherwise.
-         */
-        boolean is3xx(){
-            return statusCode >= 300 && statusCode < 400;
-        }
-
-        /**
-         * Tells whether the request yielded a 4xx status code.
-         *
-         * @return true if the request yielded a 4xx status code, false otherwise.
-         */
-        boolean is4xx(){
-            return statusCode >= 400 && statusCode < 500;
-        }
-
-        /**
-         * Tells whether the request yielded a 5xx status code.
-         *
-         * @return true if the request yielded a 5xx status code, false otherwise.
-         */
-        boolean is5xx(){
-            return statusCode >= 500 && statusCode < 600;
-        }
-
-        /**
-         * Status code getter.
-         *
-         * @return the status code.
-         */
-        int getStatusCode(){
-            return statusCode;
-        }
-
-        /**
-         * Response getter.
-         *
-         * @return the response
-         */
-        @NotNull String getResponse(){
-            return response;
-        }
-
-        @Override
-        public String toString() {
-            return "Status code: " + statusCode + ", response: " + response;
-        }
-    }
 
 
     /**
@@ -302,9 +118,9 @@ public class Trim{
         /**
          * Called when the report about an individual endpoint has been completed.
          *
-         * @param endpoint the endpoint whose report has been complete.
+         * @param model the model whose report has been complete.
          * @param completed the number of endpoints whose reports have been completed.
          */
-        void onEndpointReportComplete(@NotNull EndpointLegacy endpoint, int completed);
+        void onEndpointReportComplete(@NotNull Class<?> model, int completed);
     }
 }
